builder iplementado com interface

+++ src/builder/cominterface/IProdutoBuilder.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public interface IProdutoBuilder {
     IProdutoBuilder nome(String nome);
-    IProdutoBuilder preco(BigDecimal preco);
+    IProdutoBuilder preco(double preco);
     IProdutoBuilder  categoria(String categoria);
     IProdutoBuilder descricao(String descricao);
     IProdutoBuilder estoque(boolean disponivel);
